requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.1
Django==5.0.7
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-firebase-auth==1.2.1
django-timezone-field==7.0
djangorestframework==3.15.2
firebase-admin==6.5.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.7
intasend-python==1.1.2
kombu==5.3.7
msgpack==1.1.0
pika==1.3.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
